move custom ICDS UITab
